+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Pre-commit hook now validates for unused dependencies in mix.lock
+- Automatic `usage_rules` integration in igniter install process:
 
 ## [0.1.0] - 2025-07-25